removed the stabilizer unittest file and moved the code over into the test_mixer_class unittest file
